render deletion form conditionally based on authentication in $postId.tsx

+++ app/routes/posts/$postId.tsx
@@ -3,11 +3,14 @@ import type { ActionFunction, LoaderFunction } from '@remix-run/node';
 import { redirect } from '@remix-run/node';
 import { Link, useLoaderData } from '@remix-run/react';
 import { Heading } from '@chakra-ui/react';
+
 import { prisma } from '~/db';
+import { getUser } from '~/utils/session.server';
 
 import Button from '~/components/button';
 
-export const loader: LoaderFunction = async ({ params }) => {
+export const loader: LoaderFunction = async ({ params, request }) => {
+  const user = await getUser(request);
   const post = await prisma.post.findUnique({
     where: { id: params.postId },
     include: { user: { select: { username: true } } },
@@ -18,19 +21,23 @@ export const loader: LoaderFunction = async ({ params }) => {
       status: 404,
     });
 
-  return post;
+  return { post, user };
 };
 
 export const action: ActionFunction = async ({ params, request }) => {
   const form = await request.formData();
   if (form.get('_method') === 'delete') {
+    const user = await getUser(request);
     const post = await prisma.post.findUnique({
       where: { id: params.postId },
     });
 
     if (!post) throw new Error('Post not found');
 
-    await prisma.post.delete({ where: { id: params.postId } });
+    if (user && post.userId === user.id) {
+      await prisma.post.delete({ where: { id: params.postId } });
+    }
+
     return redirect('/posts');
   }
 };
@@ -49,10 +56,14 @@ export function CatchBoundary() {
 
 export default function Post() {
   const {
-    content,
-    createdAt,
-    title,
-    user: { username },
+    post: {
+      content,
+      createdAt,
+      title,
+      userId,
+      user: { username },
+    },
+    user,
   } = useLoaderData();
 
   return (
@@ -71,10 +82,12 @@ export default function Post() {
       >
         {content}
       </p>
-      <form method="POST">
-        <input type="hidden" name="_method" value="delete" />
-        <Button>Delete</Button>
-      </form>
+      {user.id === userId && (
+        <form method="POST">
+          <input type="hidden" name="_method" value="delete" />
+          <Button>Delete</Button>
+        </form>
+      )}
     </div>
   );
 }
